==> code/02_CRI_Common_Specifications.Rmd <==
---
title: "02. Common Specifications"
output: html_document
---

```{r setup, include=FALSE}
rm(list = ls())
library("pacman")

pacman::p_load("devtools","ggplot2","tidyverse","readr","ExPanDaR","plotscale","lattice","tidyr","mlogit","knitr","grid","zoo","ggpubr","ragg","factoextra","cluster","kableExtra","skimr","WeightedCluster","Hmisc")
```

## Load Data

```{r load, warning=FALSE,message=FALSE, include=FALSE}

cri <- read.csv(file = here::here("data/cri.csv"), header = TRUE)
cri_str <- read.csv(file = here::here("data/cri_str.csv"), header = TRUE)
cri_team <- read.csv(file = here::here("data/cri_team.csv"), header = TRUE)

#remove B&F study
cri <- subset(cri, u_teamid != 0)
```

## Dissimilarity Analysis

### Model Specification Count

Many of model specifications were dropped because they came from 'extra' models provided by the teams in their code and got qualitatively coded, but that were not part of their main submitted results. Therefore, some columns of the database have zero variance. 

We generate models to distinguish: 

A) 'net migration' based measures from 'stock' based measures, even when both are used to reflect a change over time, the variable definitions are slightly different.

B) 'west' from 'non-western' immigration & 'muslim' from 'non-muslim' & 'refugee'.

```{r qual_codes, warning = FALSE}
# Add 'net migration' dummy for measurement
cri <- cri %>%
    mutate(netmig = ifelse(main_IV_measurement == "Net Migration", 1, 0),
           west = ifelse(main_IV_measurement == "Western Immigrant", 1, 0),
           nonwest = ifelse(main_IV_measurement == "Non-Western Immigrant", 1, 0),
           muslim = ifelse(main_IV_measurement == "Muslim-country Immigrant", 1, 0),
           refugee = ifelse(main_IV_measurement == "Refugee", 1, 0))

# Get data with only model specs
cri_spec_only <- cri
         
cri_spec_only <- dplyr::select(cri_spec_only, -c(count:main_IV_measurement, num_countries, inv_weight, additionalinfo, Hsupport:HresultF, AME_sup_p05:MODEL_SCORE))

# convert character to numeric
cri_spec_only <- cri_spec_only %>%
  mutate(main_IV_time = as.numeric(as.factor(main_IV_time)),
         main_IV_effect = as.numeric(as.factor(main_IV_effect)),
         package = as.numeric(as.factor(package)))

# specifications with at least three teams using them
rownames(cri_spec_only) <- cri_spec_only$id
cri_spec_only_team <- aggregate(cri_spec_only, by = list(cri_spec_only$u_teamid), FUN = "mean")
# our analysis below revealed we should try a subset
cri_spec_only2 <- subset(cri_spec_only, u_teamid != 10 & u_teamid != 61 & u_teamid != 82)
cri_spec_only <- dplyr::select(cri_spec_only,-c(u_teamid, id))
cri_spec_only2 <- dplyr::select(cri_spec_only2,-c(u_teamid, id))
cri_sums <- colSums(Filter(is.numeric, cri_spec_only_team))
cri_sums <- as.data.frame(c("sum",cri_sums))
cri_sums[,1] <- as.numeric(cri_sums[,1])
cri_sums_uncommon <- subset(cri_sums, cri_sums[,1] < 3)

# remove zero variance columns
cri_spec_only <- cri_spec_only[ - as.numeric(which(apply(cri_spec_only, 2, var) == 0))]
cri_spec_only2 <- cri_spec_only2[ - as.numeric(which(apply(cri_spec_only2, 2, var) == 0))]
```

We identified `r nrow(cri_sums) - 2` model specifications in our qualitative coding of the team's submitted code. Of these, `r ((nrow(cri_sums) - 2) - length(cri_sums_uncommon[,1]))` were *common* across teams (occurred in three or more) and `r length(cri_sums_uncommon[,1])` were *idiosyncratic* specifications (occurred in two or less teams). Of these model specifications, `r nrow(cri_spec_only)` have variance (the remainder refer to specifications in models that were dropped before the final preferred models; usually those that were seen as sensitivity models rather than main models, or those that the team in the end decided against through correspondence and error checking - remember teams could change their models at any point if they desired or had new information although this was rare).

### Dissimilarity in Decision Realms

A second look at these data suggests that we may want to separately cluster on different decisions.

1) Countries and waves
2) Variables
3) Estimation, measurement, model structure
4) All at once

```{r diss_1, warning = FALSE}
# select only country and wave variables
cri_spec_only01 <- dplyr::select(cri_spec_only, w1985:TR)
# variables in the model
cri_spec_only02 <- dplyr::select(cri_spec_only, main_IV_as_control, emigration_ivC:fbXleftright)
# est & measurement (we provide two versions, one with DVs, one without)
cri_spec_only03 <- dplyr::select(cri_spec_only, netmig, west, nonwest, muslim, refugee, main_IV_time, main_IV_effect, Jobs:ChangeFlow, twowayfe:pseudo_pnl, dichotomize)
cri_spec_only03 <- cri_spec_only03 %>%
  mutate(main_IV_time_1year = car::recode(main_IV_time, "1 = 1; c(2,3,4,5,6) = 0"),
         main_IV_time_10year = car::recode(main_IV_time, "2 = 1; c(1,3,4,5,6) = 0"),
         main_IV_time_3year = car::recode(main_IV_time, "3 = 1; c(1,2,4,5,6)= 0"),
         main_IV_time_5year = car::recode(main_IV_time, "4 = 1; c(1,2,3,5,6)= 0"),
         main_IV_time_current = car::recode(main_IV_time, "5 = 1; c(1,2,3,4,6) = 0"),
         main_IV_time_perwave = car::recode(main_IV_time, "6 = 1; c(1,2,3,4,5) = 0"),
         main_IV_effect_between = car::recode(main_IV_effect, "1 = 1; c(2,3,4) =0"),
         main_IV_effect_total = car::recode(main_IV_effect, "2 = 1; c(1,3,4) = 0"),
         main_IV_effect_unclear = car::recode(main_IV_effect, "3 = 1; c(1,2,4) = 0"),
         main_IV_effect_within = car::recode(main_IV_effect, "4 = 1; c(1,2,3) = 0"))
cri_spec_only03 <- dplyr::select(cri_spec_only03, -c(main_IV_time, main_IV_effect))
cri_spec_only03_noDV <- dplyr::select(cri_spec_only03, -c(Jobs:Health))
#make a test for all three at once
cri_spec_only04 <- dplyr::select(cri_spec_only, w1985:TR, main_IV_as_control, emigration_ivC:fbXleftright, anynonlin)
cri_spec_only04 <- cbind(cri_spec_only04, cri_spec_only03)
cri_spec_only04_noDV <- cbind(cri_spec_only04, cri_spec_only03_noDV)
# find and merge identical cases

cri$identical_samples <- wcAggregateCases(cri_spec_only01)[["disaggIndex"]]
cri$identical_variables <- wcAggregateCases(cri_spec_only02)[["disaggIndex"]]
cri$identical_estmodel <- wcAggregateCases(cri_spec_only03)[["disaggIndex"]]
cri$identical_estmodel_noDV <- wcAggregateCases(cri_spec_only03_noDV)[["disaggIndex"]]
cri$identical_all <- wcAggregateCases(cri_spec_only04)[["disaggIndex"]]
cri$identical_all_noDV <- wcAggregateCases(cri_spec_only04_noDV)[["disaggIndex"]]

# need a function to mark all duplicates, not just the subsequent instances
allDuplicated <- function(vec){
  front <- duplicated(vec)
  back <- duplicated(vec, fromLast = TRUE)
  all_dup <- front + back > 0
  return(all_dup)
}
# absolutely unique cases coded to zero
cri <- cri %>%
  mutate(identical_samples = ifelse(allDuplicated(identical_samples) == FALSE, 0, identical_samples),
         identical_variables = ifelse(allDuplicated(identical_variables) == FALSE, 0, identical_variables),
         identical_estmodel = ifelse(allDuplicated(identical_estmodel) == FALSE, 0, identical_estmodel),
         identical_estmodel_noDV = ifelse(allDuplicated(identical_estmodel_noDV) == FALSE, 0, identical_estmodel_noDV),
         identical_all = ifelse(allDuplicated(identical_all) == FALSE, 0, identical_all),
         identical_all_noDV = ifelse(allDuplicated(identical_all_noDV) == FALSE, 0, identical_all))
```

In our coding of model specifications, `r length(cri$identical_all[cri$identical_all == 0])` of `r length(cri$identical_all)` models are unique. Most teams chose to analyze these variables separately, therefore when we exclude these six variables from the model specifications there are only `r length(cri$identical_all_noDV[cri$identical_all_noDV == 0])` unique models, but all repeated models were produced within individual teams; i.e., no identical models across teams were identified.

We investigated model specification as three different domains: sample selection (which countries and waves), independent variables included in the model and finally estimation, measurement strategy and equation format. 

When considering three sub-domains of specification:

1) **Sample** specifications only produces `r length(cri$identical_samples[cri$identical_samples == 0])` unique models
2) **Variable** specifications only produces `r length(cri$identical_variables[cri$identical_variables == 0])` unique models
3) **Measurement, Estimation & Equation** specifications only produces `r length(cri$identical_estmodel[cri$identical_estmodel == 0])` unique models

#### Sample Sub-Domain

```{r desc_samples, warning = FALSE}
cri_samples <- cri %>%
    subset(!is.na(AME_Z)) %>%
    subset(identical_samples != 0) %>%
    group_by(identical_samples, u_teamid) %>%
    mutate(team_sum = 1/n()) %>%
    ungroup()
cri_samples <- cri_samples %>%
    group_by(identical_samples) %>%
    dplyr::summarize(AME_Zm = round(stats::weighted.mean(AME_Z, team_sum),3),
              AME_Zsd = round(sqrt(Hmisc::wtd.var(AME_Z, team_sum)),3),
              cases = n(),
              teams = length(unique(u_teamid)),
              Hsupm = round(stats::weighted.mean(Hsup, team_sum),3))
#exclude unique cases and cases where there are less than 5 teams that used identical specifications
cri_samples <- subset(cri_samples, teams > 5)
cri_samples <- cri_samples[order(cri_samples$cases, decreasing = TRUE),]
head(cri_samples)

```
 
Identify the sample selection that goes with cases `r cri_samples$identical_samples`.
 
```{r ident_sample, warning = FALSE}
cri_samp <- cri %>%
    subset(cri$identical_samples %in% cri_samples$identical_samples)
cri_samp <- dplyr::select(cri_samp, identical_samples, w1985:TR)

cri_sampp<- aggregate(cri_samp, by = list(cri_samp$identical_samples), FUN = "mean")

head(cri_sampp)

cri_samples$identical_samples <- c("13 Richest Democracies, 1996 & 2006 Waves","11 of 13 Richest Democracies, 1996, 2006 & 2016 Waves","13 Richest Democracies, 1996, 2006 & 2016 Waves")
```

#### Variables Sub-Domain

```{r desc_variables}
cri_variables <- cri %>%
    subset(!is.na(AME_Z)) %>%
    subset(identical_variables != 0) %>%
    group_by(identical_variables, u_teamid) %>% # count unique cases per team per type
    mutate(team_sum = n()) %>%
    ungroup()
cri_variables <- cri_variables %>%
    group_by(identical_variables) %>%
    dplyr::summarize(AME_Zm = round(stats::weighted.mean(AME_Z, team_sum),3),
              AME_Zsd = round(sqrt(Hmisc::wtd.var(AME_Z, na.rm = TRUE)),3),
              cases = n(),
              teams = length(unique(u_teamid)),
              Hsupm = round(stats::weighted.mean(Hsup, team_sum),3))

cri_variables <- subset(cri_variables, teams > 5)
cri_variables <- cri_variables[order(cri_variables$cases, decreasing = TRUE),]
head(cri_variables)
```

```{r ident_variables1, warning = FALSE}
cri_var <- cri %>%
    subset(cri$identical_variables %in% cri_variables$identical_variables)

cri_var <- dplyr::select(cri_var, identical_variables, main_IV_as_control, emigration_ivC:fbXleftright)


cri_varr <- aggregate(cri_var, by = list(cri_var$identical_variables), FUN = "mean")

head(cri_varr)

cri_variables$identical_variables <- c("No country-level, and 5 socio-demographic individual-level indep. variables", "No country-level, and 5 socio-demographic individual-level indep. variables + income")
```

#### Model Components Sub-Domain

```{r desc_estmodel, warning = FALSE}
cri_estmodel <- cri %>%
    subset(!is.na(AME_Z)) %>%
    subset(identical_estmodel != 0) %>%
    group_by(identical_estmodel, u_teamid) %>% # count unique cases per team per type
    mutate(team_sum = n()) %>%
    ungroup()
cri_estmodel <- cri_estmodel %>%
    group_by(identical_estmodel) %>%
    dplyr::summarize(AME_Zm = round(stats::weighted.mean(AME_Z, team_sum),3),
              AME_Zsd = round(sqrt(Hmisc::wtd.var(AME_Z, na.rm = TRUE)),3),
              cases = n(),
              teams = length(unique(u_teamid)),
              Hsupm = round(stats::weighted.mean(Hsup, team_sum),3))

cri_estmodel <- subset(cri_estmodel, teams > 5)
cri_estmodel <- cri_estmodel[order(cri_estmodel$cases, decreasing = TRUE),]
head(cri_estmodel)
```
```{r ident_variables, warning = FALSE}
cri_estm <- cri %>%
    subset(cri$identical_estmodel %in% cri_estmodel$identical_estmodel)
cri_estm  <- dplyr::select(cri_estm , identical_estmodel, netmig, west, nonwest, muslim, refugee, main_IV_time, main_IV_effect, Jobs:ChangeFlow, twowayfe:pseudo_pnl, dichotomize)


cri_estmm  <- aggregate(cri_estm, by = list(cri_estm$identical_estmodel), FUN = "mean")

head(cri_estmm)

cri_estmodel$identical_estmodel <- c("DV: Jobs, Test var: Stock of immigrants, logistic regression, \'two-way fixed effects\'","DV: Unemployment, Test var: Stock of immigrants, logistic regression, \'two-way fixed effects\'","DV: Income Differences, Test var: Stock of immigrants, logistic regression, \'two-way fixed effects\'","DV: Old-age Care, Test var: Stock of immigrants, logistic regression, \'two-way fixed effects\'","DV: House, Test var: Stock of immigrants, logistic regression, \'two-way fixed effects\'","DV: Health, Test var: Stock of immigrants, logistic regression, \'two-way fixed effects\'","DV: Jobs, Test var: Net migration, logistic regression, \'two-way fixed effects\'","DV: Unemployment, Test var: Net migration, logistic regression, \'two-way fixed effects\'","DV: Income Differences, Test var: Net migration, logistic regression, \'two-way fixed effects\'","DV: Old-age Care, Test var: Net migration, logistic regression, \'two-way fixed effects\'","DV: House, Test var: Net migration, logistic regression, \'two-way fixed effects\'","DV: Health, Test var: Net migration, logistic regression, \'two-way fixed effects\'")
```

### Appendix Table S4. Selected Frequencies of Model Specifications out of 1,261 Models

```{r combine, warning = FALSE}

tbl_ax <- as.data.frame(matrix(nrow = 20, ncol = 6))

tbl_ax[c(1,5,8),1] <- c("Sample Specifications","Variable Specifications","Estimation/Equation Specifications")
colnames(tbl_ax) <- c("Specifications by Sub-Domain","AME (mean)","AME (sd)","Identical Models","No. of Teams","Support Hypothesis (mean)")

tbl_ax[c(2:4),] <- cri_samples
tbl_ax[c(6:7),] <- cri_variables
tbl_ax[c(9:20),] <- cri_estmodel

rm(cri_samp, cri_sampp, cri_estm, cri_estmm, cri_var, cri_varr)

write.csv(tbl_ax, file = here::here("results/TblS4.csv"))
```

## Colophon

This file is part of [https://github.com/nbreznau/CRI](https://github.com/nbreznau/CRI), the reproduction materials for [*Observing Many Researchers using the Same Data and Hypothesis Reveals a Hidden Universe of Uncertainty*](https://doi.org/10.31222/osf.io/cd5j9).

```{r colophon, echo=FALSE}
sessionInfo()
```
